code/requirements.txt
asgiref==3.4.1
Django==3.2.10
django-sendfile==0.3.11
Pillow==8.4.0
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
